fix vite cycle dep
